+++ src/sns/commands/spec/commands/neurons/mod.rs
@@ -1,0 +1,10 @@
+mod cache;
+mod refresh;
+mod root;
+mod sort;
+
+pub(in crate::sns::commands) use cache::{
+    sns_neurons_cache_command, sns_neurons_cache_list_command, sns_neurons_cache_status_command,
+};
+pub(in crate::sns::commands) use refresh::sns_neurons_refresh_command;
+pub(in crate::sns::commands) use root::sns_neurons_command;